+++ src/components/auctions/index.ts
@@ -1,4 +1,5 @@
 export { AuctionCard } from "./AuctionCard";
 export { AuctionGrid } from "./AuctionGrid";
+export { AuctionDetailView } from "./AuctionDetailView";
 export { BidHistory } from "./BidHistory";
 export { PlaceBidForm } from "./PlaceBidForm";